python_sync/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Luís Simas"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
alembic = "^1.11.2"
psycopg2 = "^2.9.7"
fastapi = "^0.101.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.20"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "api:run"
dev = "api:dev"

[tool.taskipy.tasks]
test = "pytest"

[tool.isort]
profile = "black"
